CLOTA: Fix ANR in PhoneDoctor
BZ: 88395

An ANR could occur on PhoneDoctor when a new event is detected. This is due to
the fact the check events process is made under a broadcast receiver. This
process could take a long time and could generate an ANR.

To solve the problem the checkEvents call is made under a dedicated AsyncTask
in the NotificationReceiver.

Change-Id: I9860877e8285c415b32414accd3f6152a59ff046
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/95174
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>