PhoneDoctor : Fix corrupted zipped logfile creation
BZ: 89021

It has been observed that many event zipped logfiles downloadable on crashtool
server are corrupted and unreadable.  One cause of this problem is PhoneDoctor
could generate corrupted zippped event logfiles and upload them to crashtool
server.
Indeed, if a problem occurs during an event zipped logfile generation, the file
is not deleted and remains in application cache directory. Then, in a next
event upload loop iteration, the cache directory is read and the corrupted zip
logfile uploaded.  To fix this problem, the methods used to create the event
zipped logfiles have been updated to handle in a better way possible
exceptions, to checks if generated zip file is valid and to delete it in every
unexpected cases.

Change-Id: If23eb52fce60d8e1f13bbcd54fd2908558116194
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/96798
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>