CLOTA: Fix issues on PhoneDoctor BZ creation
BZ: 92616

Since the 4.2 Android version there isn't any input cursor in any fields
of the bug report feature of PhoneDoctor. It's a little bit difficult to fill
a field without a cursor on it. A padding of 5px around entire layout of the
bug creation screen could be a great thing.
When the user launches the BZ creation activity from a screenshot, the test on
the user information is not done, and the user can report a bug without any
user information.
When the user report a bug without close the information dialog, the screen
could be refreshed and if the user pushes again the report button, a new empty
bug is created.

To solve those problems modifications below were done.
- user_information_activity.xml and activity_bugzilla_main.xml: a 5px padding
was added around the entire layout. The cursorDrawable was set to null to force
the drawing of the default cursor.
- BugzillaMainActivity.java : a test is now done on the start of the activity
to check that user information are filled.
- BugzillaSummaryActivity : a test is now done on the start of the activity to
check that all bug fields are filled. If not, the "report bug" button is
disable.

Change-Id: I826baa2a1d777998654f4053fd9a78628d907b51
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/97124
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>