PhoneDoctor : Manage duplicate dropbox events
BZ: 90338

As soon as the full dropbox conditions are reached (ie. 10% free space on /data
partition), the DropBoxManager in Android framework enables its "trim_To_Fit"
mechanism : then it replaces every logfiles contained by the dropbox directory
by a zero-length logfile to make room on /data partition.
Because each file is deleted and next replaced by a zero-length file, it could
not be detected by the crashlogd duplicate watcher.  This patch allows to
detect a dropbox event detected twice by storing in phonedoctor database the
name of the origin file (here the dropbox logfile) associated to a crash. Then
when a dropbox event is detected, the name of its origin file is checked in
order to verify this event has not been already detected.
If a event is then detected as DUPLICATE its crashlog directory is deleted.

Change-Id: I418c0547948ec9202cd14ee75f3fefac524bf0ec
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/99528
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>