CLOTA: Fix events date display problem
BZ: 98111

The date of an event which is displayed on CrashTool it's not the real date.
It depends of the timezone's device. If CrashTool timezone and device timezone
are the same there is no problem.

Since the event date is sent as a String object and not as a Date, like before,
the issue is present. This is due to the fact that before a date convertion was
done in order to set the CrashTool timezone to the date sent. This convertion
was still present and it isn't needed anymore.

This convertion was removed from Event.java .
Moreover the DateUtils class converts a Date to a String in order to send the
event date as a String object. This class uses the local Timezone to do this,
but it needs to use the GMT timezone instead. Modifications were done in
DateUtils.java in order to do that.

Change-Id: Ia33b395c1ef78570c612b461c6b642db92029cdb
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/100203
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>