crashinfo & crashreport : INVALID uploaded events management
BZ: 97605

When an event with a missing mandatory attribute (Device_Id or Build_Id) is
uploaded, the crashtool server rejects it and returns a status with INVALID
value. In this case, this event shall not be uploaded anymore and shall be
tagged as invalid in device database.  This patch contains update to performs
those modifications in crashinfo library and in Crashreport application.

Change-Id: I5c11a3fd14f7598cf11fecd6cfc683798d666b14
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/101272
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>