phonemonitor: fix weakness in power monitor
BZ: 98214

The power monitor taps directly into the Monitor base class
PrintWriter, but it did so without proper checks.
As a consequence, a null PrintWriter could be passed to the framework
BatteryStats dumpLocked method, causing a null pointer exception.
This patch does proper checks before invoking the dumpLocked method
and protects the call to dumpLocked with a lock to avoid races.

Change-Id: I0277f623d36f7eaf69cd21ba774317b50b8e93d6
Signed-off-by: Seb Haezebrouck <sebastien.haezebrouck@intel.com>
Reviewed-on: http://android.intel.com:8080/100540
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>