phonemonitor: improve exception handling in power monitor
BZ: 98415

On some rare occasions the Parcel containing Battery stats might
be corrupted, causing BatteryStatsImpl to throw a ParcelFormatException
if it detects a bad magic number.
On some other occasions, the magic number is correct, but the Parcel
is corrupted anyway, in this case the read Parcel function will throw
a readException. We just catch all exceptions in the power monitor
refresh stats function and do not dump the data if something went wrong.

Change-Id: Ib418ac293474d7c29a4f857fd6eb089cb898bafc
Signed-off-by: Seb Haezebrouck <sebastien.haezebrouck@intel.com>
Reviewed-on: http://android.intel.com:8080/100543
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>