CLOTA: Javacrash in PhoneDoctor PDStatus
BZ: 98094

There is a Javacrash in PhoneDoctor when the application try to compute the
DATA PDStatus byte. This problem occured on a baytrail device.
This is due to the fact that in baytrail there is no modem.
Indeed, DATA byte indicates if a 3G connection is available or not.
This information is computed with the help of the TYPE_MOBILE NetworkInfo.
But without modem this component seems to not be present.

To solve the problem a try/catch was inserted and surrounds the NetworkInfo
retrieval.If the NetworkInfo returned is null, the byte is set to x.

Change-Id: Ia7f57b76ed925a8512fa99f9712fb5f70cc4aefe
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/100614
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>