CLOTA: Adapt PhoneDoctor to user build
BZ: 79878

In order to add PhoneDoctor in user build some modifications have to be done.

-The "Wifi only" checkbox in the settings activity is hidden. The paremeter is
set to true in user build.
-The LogConfig is disabled in user build. The button is removed in the
PhoneDoctor menu. In BootConfigServiceReceiver.java the action on
ACTION_BOOT_COMPLETED intent is disabled for user build.
-The rain of crashes mechanism is disabled for user build:
  In CrashReport.java all BlackLister functions called are removed for user
  build.
-All files not needed for the crash parsing are deleted. In this way, the
dataReady in events table is set to true for all events.
-The aplogs background process is disabled for the user build. For this reason
only one aplog file is uploaded for an "report a bug" or "upload my logs"
request. In this way BugzillaMainActivity.java and CrashReportHome.java are
modified to remove the possibity to choose the number of aplogs to upload.

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: I7da21fc9d083dbeeeee60cdf6165d70c8d776920
Reviewed-on: http://android.intel.com:8080/99091
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>