CLOTA: FIX DataReady issue
BZ: 100923

The DataReady parameter is always set to 1 even for TOMBSTONE, JAVACRASH and
ANR events. This is due to a mistake in Event.java since the integration of
PhoneDoctor in user build.
Indeed, in Event.java a boolean test if the build is an user build or not.
For user build dataReady is always set to 1, for other builds it's set to 0 for
JAVACRASH, TOMBSTONE and ANR events.
A none is missing before the boolean. The test is inverted and dataReady is
always set to 1 for non user builds.

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: I636497664d5c4c20f632612dfbe2d82ba03f4a0c
Reviewed-on: http://android.intel.com:8080/102218
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>