Phonedoctor : Bug fixing on upload dialog window
BZ: 97385

If an user clicks on the upload dialog window OK button without clicking
first on an item to check(now, postpone or never), the default checked item
(Upload now) is not taken into account but the previous choice is.(previous
means the last time user checked an item in the dialog window)
This is due to the fact that if user clicks directly on the OK button, the
listener associated to items to check is not called and then user choice not
stored. The action performed is then the previous user choice stored.
To solve this problem, the listener on the items to check has been removed.
The user choice is now get in the listener of the OK button.

Change-Id: Ie87bcc23050259c5ffe4edb0d73fb7e2e126c93c
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/102705
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>