CLOTA: fix NullPointerException
BZ: 101202

If the user try to open a picture album in using PhoneDoctor
a NullPointerException occurs.

This is due to the fact that the mimeType is not set in the intent which
lunches PhoneDoctor. The getType() method used by PhoneDoctor get a Null
pointer that can't be used.

To solve the problem, the resolveType() method was used instead of getType().

Change-Id: I600d334e4a8fa6e7b303e2405d70d9e5ee0efd37
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/102714
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>