CLOTA: No log in enhancement BZ
BZ: 97997

BZ which have "enhancement" severity don't need aplogs.
Indeed they deal with an enhancement and not with an issue.
There is no interest to have aplogs attached with.

When the user selects the severity "enhancement" in the dropbox list,
RadioButtons which displayed the number of aplogs to attach are removed.
In the bz_trigger file APLOG=0 is written to indicate to crashlogd that
no aplog are needed.

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: Iad7b3ab47f5c4d23c55fbfdc4064d2889185bfb7
Reviewed-on: http://android.intel.com:8080/104025
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>