CLOTA: Add date creation info in BZ history
BZ: 97993

In BZ history there is no information about when a BZ was created and so when
issue associated with occured. Testers don't always remember to indicate when
issue happened.
It may be nice to have a way to specify how long ago issue occured just
happened, within last 30 min, within last hour, within last day, etc.

Now the user can select when the issue occurs in a list (last seconds,
last minute,30 minutes,hour,day,week,month,year). It's added to the BZ
description.

For each BZ in the BZ history, there is a textView which displays the BZ's
creation date.

Change-Id: I5212d35a323d8e765b8727794d408cc2404b10c0
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/104169
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>