PhoneDoctor: Remove DropBoxManger.isFull() compile dependency
BZ: 103701

PhoneDoctor compilation fail if android.os.DropBoxManager.isFull() is
not available.

DropBoxManager.isFull() was introduced by Intel with BZ 58848, patch
http://android.intel.com:8080/#/c/71209/.

As this API is not upstreamed, when it is not available, due to no
access to sources, PhoneDoctor fails at compile time.

To prevent this, Java reflection is used at runtime to determine if
DropBoxManager.isFull() is available. If not, status is saved in PD
preferences in order to not test method availability each time.

Change-Id: If79fe2364eb5ef893353887e7400cc9a5366916a
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/104814
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>