PhoneDoctor : Logconfig : fix lost setting for default configs
BZ: 93798

When a logconfig with 'DEFAULT' type is set to OFF state by user, this state is
reset to ON after a reboot.
This problem is due to the fact this logconfig with 'DEFAULT' type is seen as
a new logconfig and then set to ON state (since its 'DEFAULT' type).
To fix this bug, the state of logconfigs set to OFF are now also saved into
SharedPreferences.  Then it's possible to know if a logconfig is really new or
not.
In the same time, two others bugs have been fixed :
- when a logconfig is 'LOCK' typed, it is now applied at 1st boot.
- when user scrolls the listView in logConfig menu, the logconfig to ON state
  are no more re-applied.

Change-Id: I4ddae9616cd0cad6b875fa77decab29adf993a22
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/104811
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>