CLOTA: Keep BZ history after FOTA
BZ: 97800

After a software update all events present in database are removed.
It could be interesting to keep all BZ created on the phone in database even
after a software update.

This patch corrects this problem.
BZ are not deleted after a software update.
BZ repository are now deleted after the BZ upload.

Change-Id: I96a5ebb1d13d12a72a817bbc213fd10e1e66b6e2
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/105070
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>