[CLOTA] GCM Step1: Receive simple text message
BZ: 117102

In order to give information to device users or to ask them to do some action
a new communication has to be set up between CrashTool and PhoneDoctor.
This communication protocol uses Google Cloud Messaging which allows a server
to send messages to different devices.
Different kind of messages can be sent to a device:
text: simple text message
app: request the user to launch or install an application
url: an url link
phone_doctor: PhoneDoctor's action to execute

The step 1 of this new implementation is just to give the possibility
to CrashTool to send a simple text message to different devices.
This message will be received on each devices and that will display
a notification on the device. This notification contains the title
and the text of the message. When the user clicks on this notification the list
of all messages received on the device is displayed. The user can choose to
hide messages that have been already acquitted.
The user can also start or stop the GCM service if he doesn't want to receive
messages.

Change-Id: I6d87e7b4374b2c56187d63782acbc74d81a211ed
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-on: http://android.intel.com:8080/106610
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>