CLOTA: Fix Javacrash on Pause Activity
BZ: 103877

Sometimes when the back touch is pressed under the StartServiceActivity,
the activity has some difficulty to unbind the CrashReportService.

On reason of the issue is that it's possible to create multiple instances of
the activity. The limit instance number is 2 and when the limit is reached no
new instance is created and no service is started. When the activity is paused
the service isn't available and the doUnbind method fails.
Another reason is that the service could be closed before the activity. For
this reason when the activity is closed it can't unbind the service because it
doesn't exist anymore.

To solve the problem the flag ACTIVITY_CLEAR_TOP was added to all intents that
launch StartServiceActivity. That avoids to have several instances of the
activity.
Moreover a check was done on the CrashReportService's state before to call
the doUnbind function.

Change-Id: I07ca698b75e7e8a5776b08b269cacd1961c7d417
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/106643
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>