CLOTA: Rework PhoneDoctor check events service
BZ: 104626

PhoneDoctor uses one Service which checks the presence of new events and then
uploads events and logs if there is data to send.
Each time an event occurs on a board an intent is sent to PhoneDoctor.
When PhoneDoctor receives this kind of intent, it starts the CrashReportService
if it's not started else it calls checkEvents.
checkEvents is the method which is responsible of the insertion of new events
in the database. To do that it reads the history_event file and for each line
it parses the line and checks if the event is already present in the database
or not.
To avoid a disynchronism of this process the history_event line process and
event creation process associated with are synchronized.
This mechanism created several ANR in the CrashReportService. This is due to
the fact that several threads could be blocked in the synchronized part and no
body released the blocked part.

To avoid this problem a new Service was created.
CheckEventsService is the service which becomes responsible of the
history_event file parse. It does the exactly same work that checkEvents
but without the synchronized part. When an intent is received by PhoneDoctor,
a request is added in a dedicated queue. Then CheckEventsService is launched
if it's not already started.
CheckEventsService reads the requests queue.
If there is request in the queue it instacks them and parses the history_event
file. Then if there is events to upload it starts CrashReportService if it's
not already started. While there is request in the queue the service loops
in this way. When the service finishes it checks that the request queue is
empty. If it's not the case it schedules his own start in 2 minutes.
CrashReportService was modified. Now it directly starts by checking the
presence of events to upload. It is only responsible of the events and logs
upload.

Change-Id: Ie2330d1f5a83aaa4ab5eda9a5912af6eeb1fa5af
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/107572
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>