CLOTA: Rework PhoneDoctor for reference phones
BZ: 111208

In order to add PhoneDoctor in reference phones, some modifications have to be
done on PhoneDoctor.
Indeed, the only feature needed on reference phones is the "report a bug"
feature. Some parts of PhoneDoctor have to be removed and other have to be
kept.

To do that some parts of the code have to be common and other have to be
specific for each application(intel phone and reference phones).
common packages and specific packages have been created in each part of the
application were the code has to be separated.
Common classes and code are put in common packages and dedicated classes or
code are put in specific packages. New classes were created in order to create
a common part of code of a class. Then dedicated part are put in an
inherited class. For example a GeneralCrashReportHome class was added to
represent the common menu of PhoneDoctor and CrashReportHome was added to a
specific package of com.intel.crashreport to put dedicated code for the intel
application or reference phone application.

Change-Id: Ie684c6d10d1bce3e1ced45ed9659dd12fbc81ecf
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/109998
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>