CLOTA: Fix null pointer exception on CrashReportService
BZ: 112854

A null pointer exception was occured in StartServiceActivity.java on
doActionAfterSelectUploadState() function.
This seems to be due to the fact that CrashReportService has been unexpectedly
closed or stopped but without close its connection to StartServiceActivity.
Thus the report management dialog is not closed and when the user select
an option the sendMessage() call made on CrashReportService throws an Exception
because the service doesn't exist anymore.

To solve the problem some modifications were done on StartServiceActivity.
Each call on CrashReportService are surrounded by a check condition to test
that the service exists.
In CrashReport a new beahvior was added on setServiceStarted().
This modification close the report management dialog possibly present in
StartServiceActivity.

Change-Id: I23dfeb46f94c489bdc9aca259be87b83dfd21855
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/112288
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>