CLOTA: Phone doctor management jump out too many times
BZ: 113682

In StartServiceActivity a dialog is displayed when there are events to upload.
This dialog is displayed when the user has chosen to be requested before to
upload events.
An inconvenience of this dialog is that when the user changes the orientation
of his board the dialog will be displayed again even if the user canceled it
before.

To solve this problem the first user choice is stored.
When StartServiceActivity is drawn (orientation change, first time display)
a check is done to get the origin of the refresh request. This is done with
the help of the intent extra fromOutside which is set to true when
'Check Events' button is pressed or when CrashReport notification is selected.
At first time display the dialog is displayed, then the user choice is stored
in dialog_value. At each time StartServiceActivity is refreshed dialog_value is
is read and the associated action is done.
cancel: a button is displayed to lets to the user the possibility to upload his
events again.
always upload: upload all events.
upload once: a button is displayed to lets the user the possibility to upload
his events again.
never or postpone: nothing to do.

Change-Id: I66518e57ce593ecf3f7a30e6bff40052629abbbe
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/114687
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>