CLOTA: No more ERROR/INFO events are reported
BZ: 118100

Since the rework of PhoneDoctor, no more ERROR or INFO events are detected
on the device.
This is due to the fact that a new enum GeneralEventGenerator was
introduced in order to split code in 2 parts: GeneralEventGenerator for common
part and EventGenerator for dedicated part (reference phone and Intel phone).
A call to setContext is missing for GeneralEventGenerator and all call to
generateEvents fail because no context are set and the database can't be
accessible.

A call to setContext was introduced in CrashReport for GeneralEventGenerator.

Change-Id: I5e3eb48cdf0c54bb4b61ae76fc870318156f0a43
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/114862
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>