phonemonitor: Fix a java crash in master service
BZ: 118524

If the master service received the start intent twice in a row,
it would crash, trying to start twice the same thread.
This could happen if phoneDoctor crashes and tried twice to restart
phonemonitor without stopping it in between.
Fix this by storing the service state in a boolean, and checking the thread state
before starting it. This avoids registering the receivers twice also.

Signed-off-by: Seb Haezebrouck <sebastien.haezebrouck@intel.com>

Change-Id: Ia0a7639fa6f40b5b7f0f546541507cb165fa60e1
Reviewed-on: http://android.intel.com:8080/115158
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Haezebrouck, Sebastien <sebastien.haezebrouck@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>