CLOTA: Notify non critical crashes
BZ: 119864

When events upload is disabled or set to always report, PhoneDoctor doesn't
notify users that an event occurs even crashes. Only critical events are
notified.
It could be a good thing to have a crash notification. Like that the user could
be always aware of what happened on a board.

A new preference was added in the preference page in order to enable or disable
crashes notification.
When an event occurs and if the crash notification is enabled, a notification
is displayed to inform the user that a crash occured.
When the user click on the notification NotifyEventActivity is displayed and
shows a summary of crashes and critical events occured on the board.

Change-Id: Iae6b12d671c2069aba5a823f2f4eaa67c24237ad
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/117057
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>