CLOTA: Add notification for all events in AskForUpload mode
BZ: 124453

Since the remove of notification for all non crash events, user which selects
AskForUpload option to upload their events are never notified that there are
data to upload excepted if crashes occured. Then it's possible that the user
never uploads his data.

To solve this issue, the notification for all events was added for AskForUpload
mode.

Change-Id: Ia8d915e9e515b4420a8ed1b6569b7f079133877e
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/119710
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>