CLOTA: Add a table Device
BZ: 123634

In order to add a new command getDevice in crashinfo, a new table device has
to be created on PhoneDoctor.
This new table contains all information about the device : device id, imei, ssn
GCM token and spid.
This information are set and update each time an event occurs.
Moreover each time the user enable/disable the GCM reception the GCM token is
updated in the device table.

Change-Id: Ia9aea854e0fe166c9e99517edc46831373c76da0
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/120256
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>