logconfig: do not set property if value does not change
BZ: 125760

Setting the logging properties will restart logcat even if the
new value is the same than the current one. This leads to
duplicate logging at phone start (as the property is persistent,
in most of the cases the values will be similar).

This fix the issue by checking that old and new values are
different before setting the property value.

Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Change-Id: Ie09875791ec3ad49850cee1b0a00d822d9c44517
Reviewed-on: http://android.intel.com:8080/120351
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>