CLOTA: Force to connect to Wifi or use MPTA for too big data
BZ: 105932

Logs attached to an event are limited to a max size of 10MB.
This is due to the fact that an user could not use a WiFi network to upload his
data and that could take a lot of time to send a such amount of data.
The limit of 10MB is an issue for users which need to have more data to
understand a crash.

To solve the problem this limitation size was removed.
If there are data to upload that have a higher size than 10MB and if there is
no wifi connection a notification is displayed to inform user that he has to
use a WiFi network or MPTA to upload his data.

Change-Id: I991db6b8216b9b04029fafb82142d890b7fa5f09
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/122473
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>