CLOTA: Fix NullPointerException in sendLogsFile
BZ: 127466

There is an issue on PhoneDoctor when the device looses his connection.
Indeed, it's possible that PhoneDoctor is uploading data on CrashTool when
the connection is lost. In this case the outputStream used to upload data
becomes null and when PhoneDoctor try to use it a NullPointerException occur

To solve the issue a check was done on the outputStream. Moreover when the
connection changes and if there is any connection available the upload proce
is cancelled.

Change-Id: I5f0f4df62941b42975962713abbf2e4f681d90dd
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/123503
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>