CLOTA: PhoneDoctor issue
BZ: 129382

There is an issue in PhoneDoctor due to PhoneInspector. Indeed, during a fast
rain of crash, PhoneDoctor received a lot of new dropbox event which launch
PhoneInspector. This could take a lot of process and could freeze the device.

To avoid this problem, a check is done when a new dropbox event occurs.
If PhoneInpsector isn't run, PhoneDoctor launches it. Else nothing is done.

Moreover, a READ_PHONE_STATE persmission was added to read the IMEI.

Change-Id: Ibf80789d5c206a616ff044674996ae03e1057c53
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/124385
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>