CLOTA: Fix Javacrash on pauseActivity
BZ: 119897

Javacrash occured on PhoneDoctor during the pause of StartServiceActivity.
This is due the fact that PhoneDoctor tries to unregister a receiver that
could not be yet registered.

To solve the problem a check was added in the unregister method to be sure
that the receiver is registered.
Moreover another check was added on CrashReportService in Destroy method to
be sure that the service is started before to use setServiceStarted(false).

Change-Id: I6a3e228b392be5eaa0e0a1aeb64d662437a3c6ff
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/124777
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>