CLOTA: Fix NullPointerException on WiFi NetworkInfo
BZ: 133849

Where no Wifi card is present on a device the Wifi NetworkInfo is null and
PhoneDoctor fails to compute the PDStatus due to a NullPointerException.

A check on the existence of the WiFi NetworkInfo was added to avoid this issue.
If the NetworkInfo is null the PDStatus is set to x.

Change-Id: I27975d5143cd362f0caf8342ba0e183e74fdd947
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/128035
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>